\name{bpp-package}
\alias{bpp-package}
\alias{pts}
\alias{ddcp}
\docType{package}
\title{Tools for Computation of Bayesian Predictive Power for a Normally Distributed Endpoint with Known Variance}
\description{Implements functions to update Bayesian Predictive Power Computations after not stopping a clinical trial at an interim analysis, whether blinded or unblinded, for a Normally distributed endpoint with known variance, with a prominent example being the hazard ratio.}
\details{
\tabular{ll}{
Package:  \tab bpp  \cr
Type:     \tab Package    \cr
Version:  \tab 1.0.0      \cr
Date:     \tab 2016-12-13 \cr
License:  \tab GPL (>=2)  \cr
LazyLoad: \tab yes        \cr
}
}

\references{
Rufibach, K., Jordan, P., Abt, M. (2016a). 
Sequentially Updating the Likelihood of Success of a Phase 3 Pivotal Time-to-Event Trial based on Interim Analyses or External Information. 
\emph{J. Biopharm. Stat.}, \bold{26}(2), 191--201.

Rufibach, K., Burger, H.U., Abt, M. (2016b). 
Bayesian Predictive Power: Choice of Prior and some Recommendations for its Use as Probability of Success in Drug Development.
\emph{Pharm. Stat.}, \bold{15}, 438--446.
}

\author{
Kaspar Rufibach (maintainer) \cr 
\email{kaspar.rufibach@roche.com}
}

\examples{

# type ?bpp_1interim for code of all the computations in Rufibach et al (2016a).

}



